Feat: Multi-GPU 코드 추가 및 validation step에서 prediction 결과를 저장하도록 코드 변경
* Multi-gpu 학습을 위해 pytorch Distributed Data Parallel(DDP) 코드 추가
* 기존의 valid epoch 후 다시 valid set을 돌면서 prediction 결과를 저장하는 코드를 valid 과정에서 prediction을 같이 저장하도록 코드 수정 (속도 개선)
* Valid step에서 바로 저장하기 위해 data loader에서 valid image path도 같이 반환하도록 코드 수정
* Geodesic distance 관련 코드 분리 및 기타 코드 수정
