+++ data_loaders/brats2021_3d.py
@@ -2,6 +2,7 @@ import os
 import random
 import torchio as tio
 from torch.utils.data import Dataset, DataLoader
+from torch.utils.data.distributed import DistributedSampler
 
 from data_loaders.transforms import get_transform
 
@@ -24,8 +25,8 @@ class BratsDataset(Dataset):
         if self.transform:
             subject = self.transform(subject)
 
-        # Shape : (C, W, H, D), (W, H, D)
-        return subject.image.data, subject.label.data[0, ...]
+        # Shape : [list], (C, W, H, D), (W, H, D)
+        return self.image_paths[index], subject.image.data, subject.label.data[0, ...]
 
     def _set_file_paths(self):
         self.image_paths = []
@@ -43,24 +44,36 @@ def get_dataloaders(config):
     
     random.shuffle(train_valid_files)
     split_idx = int(len(train_valid_files) * 0.8)
-    train_paths = [os.path.join(train_dir, f) for f in train_valid_files[:split_idx]]
-    valid_paths = [os.path.join(train_dir, f) for f in train_valid_files[split_idx:]]
+    train_paths = sorted([os.path.join(train_dir, f) for f in train_valid_files[:split_idx]])
+    valid_paths = sorted([os.path.join(train_dir, f) for f in train_valid_files[split_idx:]])
 
     dataloaders = {}
     for split in ["train", "valid"]:
         if split == "train":
             train_dataset = BratsDataset(paths=train_paths, transform=get_transform(split))
+            if config.exp.multi_gpu:
+                train_sampler = DistributedSampler(train_dataset)
+                shuffle = False
+            else:
+                train_sampler = None
+                shuffle = True
             dataloaders[split] = DataLoader(
-                train_dataset,
+                dataset=train_dataset,
+                sampler=train_sampler,
                 batch_size=config.data.batch_size, 
-                shuffle=True, 
+                shuffle=shuffle, 
                 num_workers=config.data.num_workers,
                 pin_memory=True
             )
         elif split == "valid":
             valid_dataset = BratsDataset(paths=valid_paths, transform=get_transform(split))
+            if config.exp.multi_gpu:
+                valid_sampler = DistributedSampler(valid_dataset)
+            else:
+                valid_sampler = None
             dataloaders[split] = DataLoader(
-                valid_dataset,
+                dataset=valid_dataset,
+                sampler=valid_sampler,
                 batch_size=config.data.batch_size, 
                 shuffle=False, 
                 num_workers=config.data.num_workers,
